+++ Readme.md
@@ -1,26 +1,16 @@
-# Continuous-in-Depth Neural Networks
+# Stateful ODE-Nets
 
 Introduction (todo).
 
-This directory contains the implementation of ContinuousNet to accompany the preprint
-> [Queiruga, Alejandro F., N. Benjamin Erichson, Dane Taylor and Michael W. Mahoney. “Continuous-in-Depth Neural Networks.” ArXiv abs/2008.02389 (2020)](https://arxiv.org/abs/2008.02389)
-
-<img src="https://github.com/erichson/data/blob/master/img/ContinuousNet_overview.png" width="750">
 
 
 ## Get Started
 
 Just clone the ContinuousNet repository to your local system and you are ready to go:
 ```
-git clone https://github.com/afqueiruga/ContinuousNet
+git clone https://github.com/afqueiruga/StatefulOdeNets
 ```
 
-Also, you need a fork of `torchdiffeq` that is slightly modified in order to expose additional solver options: 
-```
-pip install git+https://github.com/afqueiruga/torchdiffeq
-```
-
-Note, this implementation best works with [PyTorch 1.6](https://pytorch.org/).
 
 ## Training
 
@@ -66,46 +56,10 @@ After training the model checkpoint is saved in a folder called results.
 (todo)
 
 ```
-python train_resnet.py --name cifar10 --epochs 120 --arch ResNet --lr_decay_epoch 30 60 90 --depth_res 20
-```
-
-```
-python3 cli.py --model ContinuousNet --scheme euler --dataset CIFAR10 --epochs 120  --lr_decay_epoch 30 60 90 --initial_time_d 2
-```
-
-```
-python3 cli.py --model ContinuousNet --scheme rk4_classic --dataset CIFAR10 --epochs 120  --lr_decay_epoch 30 60 90 --initial_time_d 2 --weight_decay 1e-4
-```
 
 
 | Model           |  Units           | Refined     | Scheme      | #parms  | Test Accuracy | Time |
 | ----------------|:----------------:|:----------: |:----------: |:-------:|:-------------:|:----:|
-| ResNet-20 (v2)  | 2-2-2            |  -          | -           | 0.27M   | 91.31%        |48 (m)|
-| ContinuousNet   | 2-2-2            |  -          | Euler       | 0.27M   | 91.41%        |20 (m)|
-| ContinuousNet   | 2-2-2            |  -          | RK4-classic | 0.27M   | 91.01%        |50 (m)|
-| ContinuousNet   | 1-1-1 -> 2-2-2   | 25          | RK4-classic | 0.27M   | 91.09%        |47 (m)|
-| ContinuousNet   | 1-1-1 -> 2-2-2   | 25          | Midpoint    | 0.27M   | 90.67%        |29 (m)|
-
-
-
-| Model           |  Units          |Refined     | Scheme      | #parms  | Test Accuracy | Time  |
-| ----------------|:---------------:|:----------:|:----------: |:-------:|:-------------:|:-----:|
-| ResNet-52 (v2)  | 8-8-8           | -          | -           | 0.85M   | 93.11%        |105 (m)|
-| ContinuousNet   | 8-8-8           | -          | Euler       | 0.86M   | 93.11%        | 83 (m)|
-| ContinuousNet   | 8-8-8           | -          | RK4-classic | 0.86M   | 93.29%        |279 (m)|
-| ContinuousNet   | 1-1-1 -> 8-8-8  | 30, 50, 70 | RK4-classic | 0.86M   | 93.06%        |199 (m)|
-
-
-
-## Examples and Performance on CIFAR-100
-
-(todo)
-
-| Model             |  Units  | Scheme      | #parms  | Test Accuracy | Time  |
-| ------------------|:-------:|:----------: |:-------:|:-------------:|:-----:|
-| WideResNet-58     | 8-8-8   | -           |  13.63M |   79.16%      |193 (m)|
-| WideContinuousNet | 8-8-8   | Euler       |  13.63M |   78.45%      |159 (m)|
-
 
 
 ## Evaluation
