--- residual_net/__init__.py
@@ -1,2 +1,0 @@
-from .ResNetv2 import ResNet as ResNetv2
-from .WideResNet import WideResNet
