+++ run_cifar10.py
@@ -1,12 +1,9 @@
 from typing import Tuple
 import sys
 
-from continuous_net_jax.run_experiment import run_an_experiment
+from continuous_net.run_experiment import run_an_experiment
 
 import argparse
-
-# Example of how to run:
-# export CUDA_VISIBLE_DEVICES=6; python run_cifar10_jax_version.py --alpha 1 --seed 1 --refine_epochs 20 30 40 50 60
 
 
 parser = argparse.ArgumentParser(description='training parameters')
@@ -15,7 +12,7 @@ parser.add_argument('--wd', type=float, default=5e-4, help='weight decay paramet
 parser.add_argument('--scheme', type=str, default="Euler", help='numerical integrator scheme')
 parser.add_argument('--epsilon', type=float, default=1.0, help='epsilon')
 parser.add_argument('--basis', type=str, default="piecewise_constant", help='basis function')
-parser.add_argument('--alpha', type=int, default=1, help='multiplier on number of channels = 16*alpha')
+parser.add_argument('--alpha', type=int, default=1, help='widht of network')
 parser.add_argument('--n_steps', type=int, default=1, help='number of steps')
 parser.add_argument('--n_basis', type=int, default=1, help='number of basis functions')
 parser.add_argument('--epochs', type=int, default=180, help='number of epochs')
@@ -31,15 +28,17 @@ parser.add_argument('--which_model', type=str, default="ContinuousNet", help='wh
 args = parser.parse_args()
 print(args)
 
+root = './'
+
 run_an_experiment(
           dataset_name='CIFAR10',
           save_dir=args.save_dir,
           which_model=args.which_model,
           alpha=args.alpha, 
-          hidden=8, 
+          hidden=16, 
           n_step=args.n_steps, 
           n_basis=args.n_basis, 
-          norm="BatchNorm",
+          norm="BatchNorm-opt-flax",
           basis=args.basis,
           scheme=args.scheme,
           epsilon=args.epsilon,
